Changes in the evaluation function
